feat: update version to 2.2.4 and add new components for rich text editing and API token management
- Added RichTextPromptModal.vue for rich text input prompts.
- Added RichTextTableMenu.vue for table manipulation options.
- Introduced ApiTokensCard.vue for managing API tokens with create, revoke, and display functionalities.
- Implemented ChangePasswordModal.vue for user password updates with validation.
