--- app/components/common/BrandingHeader.vue
@@ -1,27 +1,0 @@
-<template>
-  <div class="flex items-center gap-3 p-3 border-b border-[var(--border-default)]">
-    
-    <div class="relative group">
-      <div
-        class="w-10 h-10 rounded-[var(--radius-control)] [background-image:var(--gradient-primary)] flex items-center justify-center shadow-lg"
-      >
-        <span class="text-white font-bold text-lg">E</span>
-      </div>
-
-      <div
-        class="absolute inset-0 w-10 h-10 rounded-[var(--radius-control)] [background-image:var(--gradient-primary)] opacity-50 blur-md -z-10"
-      ></div>
-    </div>
-
-    <div class="flex flex-col">
-      <h1
-        class="font-bold text-xl [background-image:var(--gradient-primary)] bg-clip-text text-transparent"
-      >
-        Enfyra
-      </h1>
-      <p class="text-xs text-muted-foreground font-medium tracking-wide">
-        Content Management
-      </p>
-    </div>
-  </div>
-</template>
