+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "enfyra-app",
-  "version": "2.2.3",
+  "version": "2.2.4",
   "description": "Dynamic admin interface for Enfyra backend - auto-generates management UI from database schemas",
   "license": "MIT",
   "author": "dothinh115 <dothinh115@gmail.com>",
